+++ kompendium-auth/src/main/kotlin/io/bkbn/kompendium/auth/KompendiumAuth.kt
@@ -1,17 +1,16 @@
 package io.bkbn.kompendium.auth
 
-import io.ktor.auth.Authentication
-import io.ktor.auth.basic
-import io.ktor.auth.BasicAuthenticationProvider
-import io.ktor.auth.jwt.jwt
-import io.ktor.auth.jwt.JWTAuthenticationProvider
 import io.bkbn.kompendium.core.Kompendium
 import io.bkbn.kompendium.oas.security.ApiKeyAuth
 import io.bkbn.kompendium.oas.security.BasicAuth
 import io.bkbn.kompendium.oas.security.BearerAuth
 import io.bkbn.kompendium.oas.security.OAuth
-import io.ktor.auth.AuthenticationRouteSelector
+import io.ktor.auth.Authentication
+import io.ktor.auth.BasicAuthenticationProvider
 import io.ktor.auth.OAuthAuthenticationProvider
+import io.ktor.auth.basic
+import io.ktor.auth.jwt.JWTAuthenticationProvider
+import io.ktor.auth.jwt.jwt
 import io.ktor.auth.oauth
 
 /**
@@ -21,12 +20,6 @@ import io.ktor.auth.oauth
 object KompendiumAuth {
 
   private const val DEFAULT_NAME = "default"
-
-  init {
-    Kompendium.addCustomRouteHandler(AuthenticationRouteSelector::class) { route, tail ->
-      calculate(route.parent, tail)
-    }
-  }
 
   /**
    * Adds a basic auth schema to the [Kompendium] security schemas list with the specified name