+++ kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/Kompendium.kt
@@ -2,8 +2,6 @@ package io.bkbn.kompendium.core
 
 import io.bkbn.kompendium.core.metadata.ErrorMap
 import io.bkbn.kompendium.core.metadata.SchemaMap
-import io.bkbn.kompendium.core.path.IPathCalculator
-import io.bkbn.kompendium.core.path.PathCalculator
 import io.bkbn.kompendium.oas.OpenApiSpec
 import io.bkbn.kompendium.oas.info.Info
 import io.bkbn.kompendium.oas.schema.TypedSchema
@@ -25,8 +23,6 @@ object Kompendium {
     paths = mutableMapOf()
   )
 
-  fun calculatePath(route: Route) = PathCalculator.calculate(route)
-
   fun resetSchema() {
     openApiSpec = OpenApiSpec(
       info = Info(),
@@ -38,12 +34,5 @@ object Kompendium {
 
   fun addCustomTypeSchema(clazz: KClass<*>, schema: TypedSchema) {
     cache = cache.plus(clazz.simpleName!! to schema)
-  }
-
-  fun <T : RouteSelector> addCustomRouteHandler(
-    selector: KClass<T>,
-    handler: IPathCalculator.(Route, String) -> String
-  ) {
-    PathCalculator.addCustomRouteHandler(selector, handler)
   }
 }