+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Bumped Kotlin to 1.6
 - Annotations now live in a separate module.  (Should not impact end users as module is imported as api dependency by core).
 - Kotest as the testing framework of choice
+- Path calculation removed in favor of built-in route toString
 
 ### Remove